changed vehicle route indexer type to size_t
added a bunch of (probably somewhat redundant) initializers to the vehicle route indexer. Should be zero after traversing input link as well.

+++ branches/busmezzo/mezzo_lib/src/route.h
@@ -45,7 +45,7 @@ class Route
 	/** @ingroup DRT 
 		@{
 	*/
-	Link* nextlink(int currentlink_idx); //!< returns nextlink based on index in links vector
+	Link* nextlink(size_t currentlink_idx); //!< returns nextlink based on index in links vector
 	bool has_cycle(); //!< returns true of this route contains repeating links
 	/**@}*/
